Move throttle utils to common file
